Edited sup_tab_6_cohort
Deleted the filter on fy_due because it seemed to be getting rid of a lot of people at the start of the cohort e.g. if you were screened in May 2019 and given annual follow up then you'd be filtered out because your new expected screening date would only be 180 days after that. I removed the whole thing as I'm not sure this filter is strictly required especially since long intervals are allowed later on.

Removed surveillance interval from the group_by because there were cases where it was double counting people who'd moved from one type of surveillance to another.

Changed the group_by from date_screen_last to date_screen since I think it makes more sense to operate with respect to the surveillance appointments as opposed to the appointment when surveillance is deemed necessary.